refactor: VoucherConsoleIO 클래스를 AbstractConsoleIO와 VoucherConsoleIO로 분리
